Fix conversions precission

+++ cosmpy/aerial/client/__init__.py
@@ -110,7 +110,7 @@ from cosmpy.tendermint.rest_client import (
     CosmosBaseTendermintRestClient as TendermintRestClient,
 )
 from cosmpy.tx.rest_client import TxRestClient
-
+from decimal import Decimal
 
 DEFAULT_QUERY_TIMEOUT_SECS = 15
 DEFAULT_QUERY_INTERVAL_SECS = 2
@@ -349,7 +349,7 @@ class LedgerClient:
         resp = self.bank.Balance(req)
         assert resp.balance.denom == denom  # sanity check
 
-        return int(float(resp.balance.amount))
+        return int(Decimal(resp.balance.amount))
 
     def query_bank_all_balances(self, address: Address) -> List[Coin]:
         """Query bank all balances.
@@ -412,7 +412,7 @@ class LedgerClient:
             validators.append(
                 Validator(
                     address=Address(validator.operator_address),
-                    tokens=int(float(validator.tokens)),
+                    tokens=int(Decimal(validator.tokens)),
                     moniker=str(validator.description.moniker),
                     status=ValidatorStatus.from_proto(validator.status),
                 )
@@ -443,14 +443,14 @@ class LedgerClient:
                 for reward in rewards_resp.rewards:
                     if reward.denom == self.network_config.staking_denomination:
                         stake_reward = (
-                            int(float(reward.amount)) // COSMOS_SDK_DEC_COIN_PRECISION
+                            int(Decimal(reward.amount)) // COSMOS_SDK_DEC_COIN_PRECISION
                         )
                         break
 
                 current_positions.append(
                     StakingPosition(
                         validator=Address(item.delegation.validator_address),
-                        amount=int(float(item.balance.amount)),
+                        amount=int(Decimal(item.balance.amount)),
                         reward=stake_reward,
                     )
                 )
@@ -464,7 +464,7 @@ class LedgerClient:
                 total_unbonding = unbonding_summary.get(validator, 0)
 
                 for entry in item.entries:
-                    total_unbonding += int(float(entry.balance))
+                    total_unbonding += int(Decimal(entry.balance))
 
                 unbonding_summary[validator] = total_unbonding
